refactor(structure): move spherical harmonics to function_basis
